Make IsMatch an IngredientSearchTerm method
