+++ src/WebClient/Pages/Recipes.razor
@@ -1,7 +1,7 @@
 ﻿@page "/recipes"
 @using System.Collections.Specialized
+@using System.Linq
 @using System.Net.Http.Json
-@using RecipeManager.ApplicationCore.Extensions
 @using RecipeManager.ApplicationCore.Models
 @using RecipeManager.ApplicationCore.Resources
 @using RecipeManager.ApplicationCore.Search
@@ -24,7 +24,7 @@ Search recipes by the ingredients they contain.
     foreach (var recipe in Results.Value)
     {
         <MatCard Class="mat-elevation-z3" Style="margin: 8px; padding: 8px;" Stroke="true">
-            <h1>@(recipe.Title + (recipe.IsAvailable ? " - Avialable" : string.Empty))</h1>
+            <h1>@(recipe.Title + (recipe.Ingredients.All(x => IngredientTerms?.Any(y => y.IsMatch(x)) ?? false) ? " - Available" : string.Empty))</h1>
             <MatTable Items="@recipe.Ingredients" ShowPaging="false">
                 <MatTableHeader>
                     <th>Name</th>
@@ -41,7 +41,7 @@ Search recipes by the ingredients they contain.
                     <td>@context.Units</td>
                      @if (IngredientTerms != null)
                      {
-                         <th>@(context.IsAvailable ? "x" : string.Empty)</th>
+                         <th>@(IngredientTerms.Any(x => x.IsMatch(context.Name, context.Quantity, context.Units)) ? "x" : string.Empty)</th>
                      }
                 </MatTableRow>
             </MatTable>
@@ -56,7 +56,7 @@ Search recipes by the ingredients they contain.
     
     bool IsSearching { get; set; }
 
-    PagedCollection<AvailableRecipe> Results { get; set; }
+    PagedCollection<RecipeResource> Results { get; set; }
 
     private ICollection<IngredientSearchTerm> IngredientTerms { get; set; }
     
@@ -113,7 +113,7 @@ Search recipes by the ingredients they contain.
         StateHasChanged();
         try
         {
-            Results = await Http.GetFromJsonAsync<PagedCollection<AvailableRecipe>>(uri);
+            Results = await Http.GetFromJsonAsync<PagedCollection<RecipeResource>>(uri);
         }
         catch (Exception ex)
         {
@@ -121,19 +121,6 @@ Search recipes by the ingredients they contain.
             IsSearching = false;
             StateHasChanged();
             return;
-        }
-
-        if (IngredientTerms != null)
-        {
-            foreach (var recipe in Results.Value)
-            {
-                foreach (var ingredient in recipe.Ingredients)
-                {
-                    ingredient.IsAvailable = IngredientTerms.Any(x => ingredient.IsMatch(x));
-                }
-                
-                recipe.IsAvailable = recipe.Ingredients.All(x => x.IsAvailable);
-            }
         }
 
         IsSearching = false;
